Refactor Challenges.jsx and create QuestionCard.jsx and QuizContainer.jsx
